abssceneを編集しますた

+++ 21GameJum 0925/Assets/Aoyama/ScoreSystem.cs
@@ -21,6 +21,6 @@ public class ScoreSystem : MonoBehaviour
     void Update()
     {
         score = player.transform.position.x - orizinPosition.position.x;
-        scoreText.text ="キョリ : " + score;
+        scoreText.text ="キョリ : " + (int)score;
     }
 }
